+++ src/main/java/com/example/minimal_prod_backend/dto/UnidadConversionInput.java
@@ -6,7 +6,7 @@ import java.math.BigDecimal;
 
 @Data
 public class UnidadConversionInput {
-    private Integer idOrigen;
-    private Integer idDestino;
+    private Long idOrigen;
+    private Long idDestino;
     private BigDecimal factor;
 }
