+++ src/main/java/com/example/minimal_prod_backend/dto/UnidadMedidaInput.java
@@ -7,7 +7,7 @@ public class UnidadMedidaInput {
     private String codigo;
     private String nombre;
     private String abreviatura;
-    private Integer idTipo;
+    private Long idTipo;
     private Boolean esActiva;
     private Boolean esBase;
 }
