+++ src/main/java/com/example/minimal_prod_backend/repository/EstadoOrdenRepository.java
@@ -3,5 +3,5 @@ package com.example.minimal_prod_backend.repository;
 import com.example.minimal_prod_backend.entity.EstadoOrden;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoOrdenRepository extends JpaRepository<EstadoOrden, Integer> {
+public interface EstadoOrdenRepository extends JpaRepository<EstadoOrden, Long> {
 }
