+++ src/main/java/com/example/minimal_prod_backend/repository/LoteProduccionRepository.java
@@ -3,5 +3,5 @@ package com.example.minimal_prod_backend.repository;
 import com.example.minimal_prod_backend.entity.LoteProduccion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoteProduccionRepository extends JpaRepository<LoteProduccion, Integer> {
+public interface LoteProduccionRepository extends JpaRepository<LoteProduccion, Long> {
 }
