+++ src/main/java/com/example/minimal_prod_backend/repository/TipoCostoRepository.java
@@ -3,5 +3,5 @@ package com.example.minimal_prod_backend.repository;
 import com.example.minimal_prod_backend.entity.TipoCosto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoCostoRepository extends JpaRepository<TipoCosto, Integer> {
+public interface TipoCostoRepository extends JpaRepository<TipoCosto, Long> {
 }
