+++ src/main/java/com/example/minimal_prod_backend/repository/UnidadMedidaRepository.java
@@ -3,5 +3,5 @@ package com.example.minimal_prod_backend.repository;
 import com.example.minimal_prod_backend.entity.UnidadMedida;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnidadMedidaRepository extends JpaRepository<UnidadMedida, Integer> {
+public interface UnidadMedidaRepository extends JpaRepository<UnidadMedida, Long> {
 }
