+++ src/main/java/com/example/minimal_prod_backend/repository/UnidadMedidaTipoRepository.java
@@ -3,5 +3,5 @@ package com.example.minimal_prod_backend.repository;
 import com.example.minimal_prod_backend.entity.UnidadMedidaTipo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UnidadMedidaTipoRepository extends JpaRepository<UnidadMedidaTipo, Integer> {
+public interface UnidadMedidaTipoRepository extends JpaRepository<UnidadMedidaTipo, Long> {
 }
