+++ src/main/java/com/example/minimal_prod_backend/service/CostoOrdenService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface CostoOrdenService {
     List<CostoOrdenResponse> getCostosOrden();
-    CostoOrdenResponse getCostoOrdenById(Integer id);
+    CostoOrdenResponse getCostoOrdenById(Long id);
     CostoOrdenResponse createCostoOrden(CostoOrdenInput costoOrdenInput);
-    CostoOrdenResponse updateCostoOrden(Integer id, CostoOrdenInput costoOrdenInput);
-    void deleteCostoOrden(Integer id);
+    CostoOrdenResponse updateCostoOrden(Long id, CostoOrdenInput costoOrdenInput);
+    void deleteCostoOrden(Long id);
 }
